refactor(minmax): rename package
